Refactor:Move ProcessingDialog to uiblocks
