+++ lib/features/hashing/uiblocks/VideoButton.dart
@@ -1,5 +1,5 @@
 import 'package:decentproof/features/hashing/interfaces/IHashingService.dart';
-import 'package:decentproof/shared/ProcessingDialog.dart';
+import 'package:decentproof/shared/uiblocks/ProcessingDialog.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
